create shared helper method to get all pokemon for a specified generation
