+++ src/web-app/pages/index.tsx
@@ -2,12 +2,11 @@ import { GetStaticProps, NextPage } from 'next';
 import Image from 'next/image';
 import { useTranslation } from 'next-i18next';
 import { serverSideTranslations } from 'next-i18next/serverSideTranslations';
-import { AlertType, calculateScore, fetchPokemonGenerations } from 'poke-guess-shared';
+import { AlertType, calculateScore, fetchPokemonGenerations, getPokemonForGeneration } from 'poke-guess-shared';
 import { IPokemonGameSave } from 'poke-guess-shared';
 import { IPokemon } from 'poke-guess-shared';
 import { IPokemonApiCache } from 'poke-guess-shared';
 import { IPokemonGeneration } from 'poke-guess-shared';
-import Pokedex, { Generation } from 'pokedex-promise-v2';
 import React, { useCallback, useEffect, useRef, useState } from 'react';
 
 import AutoDismissAlert from '../components/alerts/autoDismissAlert';
@@ -81,26 +80,7 @@ const Home: NextPage = () => {
             // If a cached generation is found and it contains the list of pokemon, return the result.
             return cachedGeneration.pokemon[languageKey];
         }
-        // Create pokedex instance.
-        const P = new Pokedex();
-        // Query the generation.
-        const generation = (await P.getGenerationByName(genName)) as Generation;
-        // Retrieve the promises for the desired language pokemon models.
-        const pokemonPromises = generation.pokemon_species.map(async (p) => {
-            const pokemonSpecies = (await P.getPokemonSpeciesByName(p.name)) as Pokedex.PokemonSpecies;
-            const nameForRequestedLanguage = pokemonSpecies.names.filter((pokeAPIName) => pokeAPIName.language.name === languageKey)[0].name;
-            const pokemonModel: IPokemon = {
-                id: pokemonSpecies.id,
-                name: nameForRequestedLanguage,
-            };
-            return pokemonModel;
-        });
-        // Wait for everything to complete.
-        const result = await Promise.all(pokemonPromises).then((result) => {
-            // Order the result.
-            result.sort((a, b) => (a.id < b.id ? -1 : 1));
-            return result;
-        });
+        const result = await getPokemonForGeneration(genName, languageKey);
         if (!cachedGeneration || (cachedGeneration && (!cachedGeneration.pokemon[languageKey] || cachedGeneration.pokemon[languageKey].length <= 0))) {
             // If there is no cache for the current generation, or the cached generation contains no pokemon, fill the cache.
             const tmpCacheModel = { ...cache };
@@ -108,14 +88,14 @@ const Home: NextPage = () => {
                 // If there are no generations ion cache yet, init it.
                 tmpCacheModel.generations = [];
             }
-            const currentGenCache = tmpCacheModel.generations?.find((g) => g.name === generation.name);
+            const currentGenCache = tmpCacheModel.generations?.find((g) => g.name === genName);
             if (currentGenCache) {
                 // If the current gen exists, add the pokemon for the requested language.
                 currentGenCache.pokemon[languageKey] = result;
             } else {
                 // Create gen model to add to cache.
                 const genToAdd: IPokemonGeneration = {
-                    name: generation.name,
+                    name: genName,
                     pokemon: {},
                 };
                 genToAdd.pokemon[languageKey] = result;
